+++ src/ui/service/miniquad/mod.rs
@@ -3,6 +3,9 @@
 //! `miniquad` Ui service.
 //
 
+#[cfg(feature = "alloc")]
+#[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "alloc")))]
+mod base;
 mod reexports;
 mod service; // MiniquadEventHandlerExt, MiniquadService
 mod window; // MiniquadWindow
@@ -13,6 +16,8 @@ crate::items! { // structural access: _mods, _all
 
     mod _mods {
         pub use super::{reexports::*, service::*, window::*};
+        #[cfg(feature = "alloc")]
+        pub use super::base::*;
     }
     pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]